+++ Items/Weapons/Ranged/SomaPrime.cs
@@ -48,7 +48,7 @@ Replaces standard bullets with High Velocity Bullets
         }
 
         // Terraria seems to really dislike high crit values in SetDefaults
-        public override void ModifyWeaponCrit(Player player, ref int crit) => crit += 30;
+        public override void ModifyWeaponCrit(Player player, ref float crit) => crit += 30;
 
         public override Vector2? HoldoutOffset() => new Vector2(-25, 0);
