+++ Items/Weapons/Rogue/ProfanedPartisan.cs
@@ -42,7 +42,7 @@ Stealth strikes spawn smaller spears to fly along side it");
         }
 
         // Terraria seems to really dislike high crit values in SetDefaults
-        public override void ModifyWeaponCrit(Player player, ref int crit) => crit += 15;
+        public override void ModifyWeaponCrit(Player player, ref float crit) => crit += 15;
 
         public override bool Shoot(Player player, EntitySource_ItemUse_WithAmmo source, Vector2 position, Vector2 velocity, int type, int damage, float knockback)
         {
